requirements.txt
selenium>=4.1.0
undetected-chromedriver>=3.1.5
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.9.3
webdriver-manager>=3.8.0
fake-useragent>=1.1.1 